Sensor_uint

+++ acc/node1/src/Sensor.h
@@ -10,7 +10,7 @@ class Sensor
 public:
     Sensor(uint8_t trigPin, uint8_t echoPin);
     ~Sensor(void);
-    [[ nodiscard ]] double getDistanceCm(void);
+    [[ nodiscard ]] uint16_t getDistanceCm(void);
 
 private:
     uint8_t trigPin;
